requirements.txt
Django==1.10.8
django-polymorphic==1.3
#psycopg2==2.7.3.2
pytz==2017.3
